requirements.txt
absl-py==2.1.0
altair==5.5.0
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.15.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
keras==3.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.4.1
mtcnn==0.1.1
namex==0.0.8
narwhals==1.27.1
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==3.10.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.6
psutil==5.9.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydeck==0.9.1
Pygments==2.15.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.2
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
streamlit==1.42.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.14.1
zipp==3.21.0
